+++ internal/DateComponentPicker/DateComponentRangePicker.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 
 import RangePicker from '../../RangePicker';
 
-import css from './DateComponentPicker.module.less';
+import componentCss from './DateComponentPicker.module.less';
 
 /**
  * {@link sandstone/internal/DataComponentPicker.DateComponentRangePicker} allows the selection of
@@ -52,6 +52,19 @@ const DateComponentRangePickerBase = kind({
 		accessibilityHint: PropTypes.string,
 
 		/**
+		 * Customizes the component by mapping the supplied collection of CSS class names to the
+		 * corresponding internal elements and states of this component.
+		 *
+		 * The following classes are supported:
+		 *
+		 * * `sizingPlaceholder` - The sizing placeholder class
+		 *
+		 * @type {Object}
+		 * @public
+		 */
+		css: PropTypes.object,
+
+		/**
 		 * The label to display below the picker
 		 *
 		 * @type {String}
@@ -78,8 +91,9 @@ const DateComponentRangePickerBase = kind({
 	},
 
 	styles: {
-		css,
-		className: 'dateComponentPicker'
+		css: componentCss,
+		className: 'dateComponentPicker',
+		publicClassNames: ['sizingPlaceholder']
 	},
 
 	computed: {
@@ -88,10 +102,11 @@ const DateComponentRangePickerBase = kind({
 		}
 	},
 
-	render: ({accessibilityHint, label, max, min, noAnimation, value, wrap, voiceLabel, ...rest}) => (
+	render: ({accessibilityHint, css, label, max, min, noAnimation, value, wrap, voiceLabel, ...rest}) => (
 		<RangePicker
 			{...rest}
 			accessibilityHint={(accessibilityHint == null) ? label : accessibilityHint}
+			css={css}
 			data-webos-voice-labels-ext={voiceLabel}
 			joined
 			max={max}