+++ tests/ui/specs/SwitchItem/SwitchItem-specs.js
@@ -23,12 +23,12 @@ describe.skip('SwitchItem', function () {
 		});
 
 		describe('5-way', function () {
-			it('should select the item when selected - [GT-21377]', function () {
+			it('should select the item when selected', function () {
 				Page.spotlightSelect();
 				expect(switchItem.isSelected).to.be.true();
 			});
 
-			it('should re-unselect the item when selected twice - [GT-21377]', function () {
+			it('should re-unselect the item when selected twice', function () {
 				Page.spotlightSelect();
 				Page.spotlightSelect();
 				expect(switchItem.isSelected).to.be.false();
@@ -47,12 +47,12 @@ describe.skip('SwitchItem', function () {
 		});
 
 		describe('pointer', function () {
-			it('should select the item when clicked', function () {
+			it('should select the item when clicked - [QWT-2527]', function () {
 				switchItem.self.click();
 				expect(switchItem.isSelected).to.be.true();
 			});
 
-			it('should re-unselect the item when clicked twice', function () {
+			it('should re-unselect the item when clicked twice - [QWT-2527]', function () {
 				switchItem.self.click();
 				switchItem.self.click();
 				expect(switchItem.isSelected).to.be.false();